@@ -1,7 +1,7 @@
 package com.matthewcannefax.menuplanner.grocery;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+
 import android.view.View;
 import android.widget.TextView;
 
