@@ -7,14 +7,14 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -23,7 +23,6 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.ListView;
 import android.widget.Spinner;
-import android.widget.Toast;
 
 import com.matthewcannefax.menuplanner.R;
 import com.matthewcannefax.menuplanner.recipe.recipeList.RecipeListActivity;
@@ -39,10 +38,7 @@ import com.matthewcannefax.menuplanner.utils.database.DataSource;
 import com.matthewcannefax.menuplanner.utils.notifications.NotificationHelper;
 
 import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
 import java.util.List;
-import java.util.TimeZone;
 
 //this activity is to display the selected MenuList
 //it has contains buttons to add a recipe to the menu and generate a grocery list
@@ -123,13 +119,6 @@ public class MenuListActivity extends AppCompatActivity {
             }
         });
 
-//        addIngredientButton.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View view) {
-//                addRecipeToMenu();
-//            }
-//        });
-
         NavDrawer.setupNavDrawerMenuButton(getSupportActionBar());
 
         ListView drawerListView = findViewById(R.id.navList);
@@ -165,39 +154,27 @@ public class MenuListActivity extends AppCompatActivity {
     }
 
     private void setFilterBTNListener(final Context mContext, Button filterBTN, final MenuListRecyclerAdapter allMenuAdapter) {
-        filterBTN.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                RecipeCategory selectedCat = (RecipeCategory) catSpinner.getSelectedItem();
-                if (selectedCat != RecipeCategory.ALL) {
-                    List<Recipe> filteredRecipes = new ArrayList<>();
+        filterBTN.setOnClickListener(view -> {
+            RecipeCategory selectedCat = (RecipeCategory) catSpinner.getSelectedItem();
+            if (selectedCat != RecipeCategory.ALL) {
+                List<Recipe> filteredRecipes = new ArrayList<>();
 
-                    for (Recipe r : mDataSource.getAllMenuRecipes()) {
-                        if (r.getCategory() == selectedCat) {
-                            filteredRecipes.add(r);
-                        }
+                for (Recipe r : mDataSource.getAllMenuRecipes()) {
+                    if (r.getCategory() == selectedCat) {
+                        filteredRecipes.add(r);
                     }
+                }
 
 //                    RecipeMenuItemAdapter filteredAdapter = new RecipeMenuItemAdapter(mContext, MenuListActivity.this, filteredRecipes, lv, catSpinner);
-                    MenuListRecyclerAdapter filteredAdapter = new MenuListRecyclerAdapter(getSupportFragmentManager(), mContext, filteredRecipes, catSpinner, mTwoPane);
-                    recyclerView.setAdapter(filteredAdapter);
-                    recyclerView.setLayoutManager(new LinearLayoutManager(mContext));
-                } else {
-                    MenuListRecyclerAdapter allAdapter = new MenuListRecyclerAdapter(getSupportFragmentManager(), getApplicationContext(), mDataSource.getAllMenuRecipes(), catSpinner, mTwoPane);
-                    recyclerView.setAdapter(allAdapter);
-                    recyclerView.setLayoutManager(new LinearLayoutManager(mContext));
-                }
+                MenuListRecyclerAdapter filteredAdapter = new MenuListRecyclerAdapter(getSupportFragmentManager(), mContext, filteredRecipes, catSpinner, mTwoPane);
+                recyclerView.setAdapter(filteredAdapter);
+                recyclerView.setLayoutManager(new LinearLayoutManager(mContext));
+            } else {
+                MenuListRecyclerAdapter allAdapter = new MenuListRecyclerAdapter(getSupportFragmentManager(), getApplicationContext(), mDataSource.getAllMenuRecipes(), catSpinner, mTwoPane);
+                recyclerView.setAdapter(allAdapter);
+                recyclerView.setLayoutManager(new LinearLayoutManager(mContext));
             }
         });
-    }
-
-    @Override
-    protected void onPostCreate(@Nullable Bundle savedInstanceState) {
-        super.onPostCreate(savedInstanceState);
-
-//        AdView mAdView = findViewById(R.id.addEditRecipeBanner);
-//
-//        AdHelper.SetupBannerAd(this, mAdView);
     }
 
     private void addRecipeToMenu(){
