@@ -2,16 +2,14 @@ package com.matthewcannefax.menuplanner.recipe.recipeList;
 
 import android.content.Context;
 import android.content.Intent;
-import android.provider.ContactsContract;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.CheckBox;
 import android.widget.ImageView;
-import android.widget.ListAdapter;
-import android.widget.Spinner;
 import android.widget.TextView;
 
 import com.matthewcannefax.menuplanner.R;
