@@ -3,7 +3,7 @@ package com.matthewcannefax.menuplanner.utils.database;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 public class DBHelper extends SQLiteOpenHelper {
 
